Reorganize Code (#1)
