+++ gherkin/model/world.py
@@ -1,0 +1,20 @@
+from dataclasses import dataclass
+from typing import Tuple, Union
+
+
+@dataclass()
+class World:
+    width: int
+    height: int
+    robot_origin: Tuple[int, int]
+    goal: Tuple[int, int]
+
+    def convert_to_display(
+            self, point: Tuple[Union[int, float], Union[int, float]]) -> Tuple[int, int]:
+        """
+        Convert a point from the robot coordinate system to the display coordinate system
+        """
+        robot_x, robot_y = point
+        offset_x, offset_y = self.robot_origin
+
+        return int(offset_x + robot_x), int(offset_y - robot_y)
