+++ gherkin/visuals/__init__.py
@@ -1,0 +1,3 @@
+from gherkin.visuals.visualizer import Visualizer
+
+__all__ = ["Visualizer"]
